refactor: 캐릭터 설정 의존성 방향 정리 (#119)
